+++ tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@tsconfig/node20/tsconfig.json"
+}
